+++ data-raw/zoologTaxonomy.R
@@ -1,6 +1,7 @@
 ## Code to include the `zoologTaxonomy` as public variable
 
-zoologTaxonomy <- read.csv2("inst/extdata/zoologTaxonomy.csv")
+zoologTaxonomy <- read.csv2("inst/extdata/zoologTaxonomy.csv",
+                            stringsAsFactors = FALSE)
 
 usethis::use_data(zoologTaxonomy, overwrite = TRUE, version = 2)